Add TypeInfo class to Tril
A helper for dealing with data types in Tril.
